First working version of config xml
* Site config xml now in place and documented.
* Command line parameters take precident over config file. Config properties have a final option that when set to true makes the property take precidence over every other source, including command line.
* By default config properties with a null value will be ignored. This can be overwritten using the final tag in order to use the config to clear a system property.
* Config is loaded whenever an accumulo or kafka datastore is instantiated.
