+++ docs/user/accumulo_datastore.rst
@@ -34,6 +34,18 @@ A script to bootstrap GeoMesa Accumulo on an Elastic Map Reduce cluster is provi
 Configuration
 -------------
 
+GeoMesa uses a site xml file to maintain system property configurations. This file can be found
+at ``conf/geomesa-site.xml`` of the Accumulo distribution. The default settings for GeoMesa are
+stored in ``conf/geomesa-default.xml``. Do not modify this file directly as it is never read,
+instead copy over the desired configurations into geomesa-site.xml.
+
+By default command line parameters will take precedent over this configuration file. If you wish
+a configuration item to always take precedence, even over command line parameters change the
+``<final>`` tag to true.
+
+By default configuration properties with empty values will not be applied, you can change this
+by marking a property as final.
+
 Zookeeper session timeout
 ~~~~~~~~~~~~~~~~~~~~~~~~~
 
@@ -43,6 +55,8 @@ for the GeoMesa Accumulo data store is exposed as the Java system property ``ins
 .. code-block:: bash
 
     export JAVA_OPTS="-Dinstance.zookeeper.timeout=10s"
+
+To preserve configuration set this property in the ``geomesa-site.xml``.
 
 Creating a Data Store
 ---------------------
