--- geomesa-tools-common/conf/default-site.xml
@@ -1,89 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?xml-stylesheet type="text/xsl" href="configuration.xsl"?>
-<!--
- Copyright (c) 2013-2016 Commonwealth Computer Research, Inc.
- All rights reserved. This program and the accompanying materials
- are made available under the terms of the Apache License, Version 2.0 which
- accompanies this distribution and is available at
- http://www.opensource.org/licenses/apache2.0.php.
--->
-
-<!-- Do not modify this file directly. Instead, copy entries that you    -->
-<!-- wish to modify from this file into geomesa-site.xml and change      -->
-<!-- them there.  If geomesa-site.xml does not already exist, create it. -->
-
-<configuration>
-    <property>
-        <name>geomesa.force.count</name>
-        <value>false</value>
-        <description></description>
-    </property>
-
-    <property>
-        <name>geomesa.query.cost.type</name>
-        <value></value>
-        <description></description>
-    </property>
-
-    <property>
-        <name>geomesa.query.timeout.millis</name>
-        <value></value>
-        <description></description>
-    </property>
-
-    <property>
-        <name>geomesa.scan.ranges.target</name>
-        <value>2000</value>
-        <description></description>
-    </property>
-
-    <property>
-        <name>geomesa.scan.ranges.batch</name>
-        <value>20000</value>
-        <description></description>
-    </property>
-
-    <property>
-        <name>geomesa.batchwriter.latency.millis</name>
-        <value>60000</value>
-        <description></description>
-    </property>
-
-    <property>
-        <name>geomesa.batchwriter.memory</name>
-        <value>52428800</value>
-        <description></description>
-    </property>
-
-    <property>
-        <name>geomesa.batchwriter.maxthreads</name>
-        <value>10</value>
-        <description></description>
-    </property>
-
-    <property>
-        <name>geomesa.feature.id-generator</name>
-        <value></value>
-        <description></description>
-    </property>
-
-    <property>
-        <name>geomesa.stats.compact.millis</name>
-        <value>3600000</value>
-        <description></description>
-    </property>
-
-    <property>
-        <name>geomesa.convert.scripts.path</name>
-        <value></value>
-        <description>A colon separated list of paths containing files to be loaded
-        onto the classpath. This is useful for getting scripts onto the classpath
-        for use by map-reduce ingest jobs.</description>
-    </property>
-
-    <property>
-        <name></name>
-        <value></value>
-        <description></description>
-    </property>
-</configuration>
