+++ geomesa-utils/pom.xml
@@ -193,6 +193,10 @@
                     <groupId>org.scala-lang.modules</groupId>
                     <artifactId>scala-parser-combinators_${scala.binary.version}</artifactId>
                 </dependency>
+                <dependency>
+                    <groupId>org.scala-lang.modules</groupId>
+                    <artifactId>scala-xml_${scala.binary.version}</artifactId>
+                </dependency>
             </dependencies>
         </profile>
     </profiles>
